add some common script file

+++ mysql/centos7-5.7.16/src/mysql.Tests.ps1
@@ -11,7 +11,7 @@ $commonPath = Join-Path -Path $here -ChildPath "\..\..\..\src\main\resources\com
 
 $envfile = Join-Path -Path (Split-Path -Path $here -Parent) -ChildPath fixtures/envforcodeexec.json -Resolve
 
-$result = . "$here\$sut" -envfile $envfile -action t codefile param0 param1 param2
+$result = . "$here\$sut" -envfile $envfile -action t codefile "param0 param1 param2"
 
 
 function Get-MysqlcnfValue {
@@ -29,7 +29,7 @@ function Get-MysqlcnfValue {
 
 Describe "code" {
     It "should return right result" {
-        $result -join "," | Should Be "param0,param1,param2"
+        $result | Should Be "param0 param1 param2"
 
     }
     It "should handle remaining parameters" {
@@ -60,7 +60,7 @@ Describe "code" {
     }
     It "should parse my.cnf" {
         $myenv = New-EnvForExec $envfile | Decorate-Env
-        get-mysqlcnfValue $myenv  "log-error"  | Should Be "/var/log/mysqld.log"
+        Get-mysqlcnfValue $myenv  "log-error"  | Should Be "/opt/mysqld-usage/mysqld.log"
     }
     It "should install mysql" {
         $myenv = New-EnvForExec $envfile | Decorate-Env
@@ -73,20 +73,28 @@ Describe "code" {
         $myenv.remoteFolder | Should Be "/easy-installer/"
         $rpms = (Get-UploadFiles -myenv $myenv | ? {$_ -match "(-server-\d+|-client-\d+|-common-\d+|-libs-\d+).*rpm$"} | Sort-Object) -join ' '
         $rpms | Should Be "/easy-installer/mysql-community-client-5.7.16-1.el7.x86_64.rpm /easy-installer/mysql-community-common-5.7.16-1.el7.x86_64.rpm /easy-installer/mysql-community-libs-5.7.16-1.el7.x86_64.rpm /easy-installer/mysql-community-server-5.7.16-1.el7.x86_64.rpm"
+        
+        # (Get-MysqlRpms $myenv) -join " " | Should Be "/easy-installer/mysql-community-client-5.7.16-1.el7.x86_64.rpm /easy-installer/mysql-community-common-5.7.16-1.el7.x86_64.rpm /easy-installer/mysql-community-libs-5.7.16-1.el7.x86_64.rpm /easy-installer/mysql-community-server-5.7.16-1.el7.x86_64.rpm"
 
         #(yum list installed | ? {$_ -match "mariadb-libs"}) -split "\s+" | Select-Object -First 1 | Should Be "mariadb-libs.x86_64"
 
-<#
+
         if (Centos7-IsServiceRunning "mysqld") {
             systemctl stop "mysqld"
-        } 
+        }
         $revRpms = Get-MysqlRpms $myenv | % {$_ -replace ".*/(.*)-[^-]+$", '$1'}
         [array]::reverse(($revRpms))
-        $revRpms | % {yum -y remove $_}
-        get-mysqlcnfValue $myenv "datadir" | Remove-Item -Recurse -Force
-        get-mysqlcnfValue $myenv "log-error" | Remove-Item -Force
+        $revRpms | % {yum -y remove $_} | Out-Null
+
+        if (get-mysqlcnfValue $myenv "datadir" | Test-Path) {
+            get-mysqlcnfValue $myenv "datadir" | Remove-Item -Recurse -Force
+        }
+        if (get-mysqlcnfValue $myenv "log-error" | Test-Path) {
+            get-mysqlcnfValue $myenv "log-error" | Remove-Item -Force
+        }
+        
+
         Install-Mysql $myenv
-#>
         Set-NewMysqlPassword $myenv "aks23A%soid" | Write-Host
         $LASTEXITCODE | Write-Host
         # | Should Be "Enter password: "
