+++ mysql/centos7-5.7.16/src/mysql.ps1
@@ -1,8 +1,9 @@
 # how to run this script. powershell -File /path/to/this/file.
 # ParamTest.ps1 - Show some parameter features
-# Param statement must be first non-comment, non-blank line in the script
-# runuser -s /bin/bash -c "/opt/hadoop/hadoop-2.7.3/bin/hdfs dfs -mkdir -p /user/hbase" hdfs
-# runuser -s /bin/bash -c "/opt/hadoop/hadoop-2.7.3/bin/hdfs dfs -chown hbase /user/hbase" hdfs
+# CREATE USER 'jeffrey'@'localhost' IDENTIFIED BY 'mypass';
+# ALTER USER 'jeffrey'@'localhost'IDENTIFIED BY 'mypass';
+# ALTER USER USER() IDENTIFIED BY 'mypass'; // change own password.
+# SET PASSWORD FOR 'jeffrey'@'localhost' = PASSWORD('mypass');
 
 Param(
     [parameter(Mandatory=$true)]$envfile,
@@ -44,21 +45,26 @@ function Install-Mysql {
 
 function Get-MysqlRpms {
     Param($myenv)
-    Get-UploadFiles -myenv $myenv | ? {$_ -match "-common-\d+.*rpm$"}
-    Get-UploadFiles -myenv $myenv | ? {$_ -match "-libs-\d+.*rpm$"}
-    Get-UploadFiles -myenv $myenv | ? {$_ -match "-client-\d+.*rpm$"}
-    Get-UploadFiles -myenv $myenv | ? {$_ -match "-server-\d+.*rpm$"}
+    Get-UploadFiles $myenv | ? {$_ -match "-common-\d+.*rpm$"}
+    Get-UploadFiles $myenv | ? {$_ -match "-libs-\d+.*rpm$"}
+    Get-UploadFiles $myenv | ? {$_ -match "-client-\d+.*rpm$"}
+    Get-UploadFiles $myenv | ? {$_ -match "-server-\d+.*rpm$"}
 }
 
 
 function Set-NewMysqlPassword {
     Param($myenv, $newpassword)
+    $rh = Get-Content $myenv.resultFile | ConvertFrom-Json
+
+    if ($fh.info -and $rh.info.initPasswordReseted) {
+        Write-Error "Initial Password already rested, skipping..."
+    }
     $mycnf = New-SectionKvFile -FilePath "/etc/my.cnf"
     $ef = $myenv.software.textfiles | Where-Object Name -EQ "change-init-pass.tcl" | Select-Object -First 1 -ExpandProperty content
-    $initpassword = (Get-Content $mycnf.getValue("[mysqld]", "log-error") | ? {$_ -match "A temporary password is generated"} | Select-Object -First 1) -replace ".*:\s*(.*?)\s*$",'$1'
+    $initpassword =   ( Get-SectionValueByKey -parsedSectionFile $mycnf -section "[mysqld]" -key "log-error" | Get-Content | ? {$_ -match "A temporary password is generated"} | Select-Object -First 1) -replace ".*:\s*(.*?)\s*$",'$1'
     Run-String -execute tclsh -content $ef -quotaParameter "$initpassword" "$newpassword" | Write-Output -OutVariable fromTcl
     if ($LASTEXITCODE -ne 0) {
-        Write-Error "$fromTcl"
+        Write-Error "change-init-pass.tcl exit with none zero. $fromTcl"
     }
 }
 
@@ -66,27 +72,78 @@ function New-MysqlUser {
     Param($myenv, $rootpassword)
 }
 
+function Enable-LogBin {
+    Param($myenv)
+    $sf = New-SectionKvFile -FilePath "/etc/my.cnf"
+    $mysqlds = "[mysqld]"
+    $serverId = Get-SectionValueByKey -parsedSectionFile $sf -section $mysqlds -key "server-id"
+
+    if ($serverId) {
+        Write-Error "logbin already enabled. skipping....."
+    }
+
+    if (! $serverId) {
+        Add-SectionKv -parsedSectionFile $sf -section $mysqlds -key "server-id"
+        $serverId = Get-SectionValueByKey -parsedSectionFile $sf -section $mysqlds -key "server-id"
+    }
+
+    if (! $serverId) {
+        Write-Error "There must exists an item with server-id=xxx format in my.cnf"
+    }
+<#
+    $serverIdInCfg = Get-BoxRoleConfig "mysqld" "server-id"
+    if ($serverIdInCfg) {
+        Add-SectionKv -parsedSectionFile $sf -section $mysqlds -key "server-id" -value $serverIdInCfg
+    } else {
+        if (("MYSQL_MASTER" -in $myenv.myRoles) -and $serverId) {
+            Add-SectionKv -parsedSectionFile $sf -section $mysqlds -key "server-id" -value $serverId
+        } else {
+            Add-SectionKv -parsedSectionFile $sf -section $mysqlds -key "server-id" -value (Get-Random)
+        }
+    }
+#>
+    $sf.writeToFile()
+    systemctl restart mysqld | Out-Null
+}
+
 function Write-ConfigFiles {
-    Param($myenv, $newpassword)
+    Param($myenv)
     $resultHash = @{}
     $resultHash.env = @{}
     $resultHash.info = @{}
-
-    if (! $newpassword) {
-        return
-    }
 
     #write my.cnf
     $myenv.software.textfiles | ? {$_.name -eq "my.cnf"} | Select-Object -First 1 -ExpandProperty content | Out-File -FilePath "/etc/my.cnf" -Encoding ascii
 
     $sf = New-SectionKvFile -FilePath "/etc/my.cnf"
-    $logError = $sf.getValue("[mysqld]", "log-error")
-    $datadir = $sf.getValue("[mysqld]", "datadir")
-    $pidFile = $sf.getValue("[mysqld]", "pid-file")
-    $socket =  $sf.getValue("[mysqld]", "socket")
+    $mysqlds = "[mysqld]"
+    $logError = Get-SectionValueByKey -parsedSectionFile $sf -section $mysqlds -key "log-error"
+    $datadir = Get-SectionValueByKey -parsedSectionFile $sf -section $mysqlds -key "datadir"
+    $pidFile = Get-SectionValueByKey -parsedSectionFile $sf -section $mysqlds -key "pid-file"
+    $socket = Get-SectionValueByKey -parsedSectionFile $sf -section $mysqlds -key "socket"
+    $binLog = Get-SectionValueByKey -parsedSectionFile $sf -section $mysqlds -key "log-bin"
+    # we cannot use this server-id value, because all box are same.
+    $serverId = Get-SectionValueByKey -parsedSectionFile $sf -section $mysqlds -key "server-id"
 
-    if ($logError) {
-        Remove-Item -Path $logError
+    # if bin-log init enabled.
+    if ($binLog) {
+        Comment-SectionKv -parsedSectionFile $sf -section $mysqlds -key "log-bin"
+        if ($serverId) {
+            Comment-SectionKv -parsedSectionFile $sf -section $mysqlds -key "server-id"
+        }
+        $sf.writeToFile()
+    }
+
+    if (! (Split-Path -Parent $datadir | Test-Path)) {
+        Split-Path -Parent $datadir | New-Directory | Out-Null
+    }
+
+    if (! (Split-Path -Parent $logError | Test-Path)) {
+        Split-Path -Parent $logError | New-Directory | Out-Null
+    }
+
+    if ($logError | Test-Path) {
+        Remove-Item -Path $logError | Out-Null
     }
 
     # start mysqld£¬ when mysql first start, It will create directory and user to run mysql.
@@ -116,11 +173,21 @@ switch ($action) {
     "install" {
         Install-Mysql $myenv
     }
+    "enableLogBin" {
+        enable-logbin $myenv
+    }
     "changePassword" {
         Set-NewMysqlPassword $myenv @remainingArguments
     }
+    "start" {
+        if (Centos7-IsServiceRunning "mysqld") {
+            systemctl stop mysqld
+        }
+    }
     "stop" {
-        systemctl stop mysqld
+        if (Centos7-IsServiceRunning mysqld) {
+            systemctl stop mysqld
+        }
     }
     "t" {
         $remainingArguments
